+++ apps/crm/crm/src/app/crm/pages/waterfall/document/document.component.html
@@ -125,5 +125,13 @@
     </div>
   </article>
 
-  <waterfall-interest-table [contractId]="contract.id"></waterfall-interest-table>
+  <ng-container *ngIf="interests$ | async as interests">
+    <article *ngIf="interests.length">
+      <h3>Interest Rate Details</h3>
+      <div class="surface">
+        <waterfall-interest-table [interests]="interests"></waterfall-interest-table>
+      </div>
+    </article>
+  </ng-container>
+  
 </ng-container>
