+++ apps/crm/crm/src/app/crm/pages/waterfall/document/document.component.scss
@@ -9,6 +9,6 @@ header {
   justify-content: space-between;
 }
 
-h3, waterfall-interest-table {
+h3 {
   margin-top: 40px;
 }
