+++ apps/crm/crm/src/app/crm/pages/waterfall/document/document.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute } from '@angular/router';
-import { Scope, Term, WaterfallContract, getDeclaredAmount, rightholderGroups } from '@blockframes/model';
+import { ConditionInterest, Scope, Term, WaterfallContract, getDeclaredAmount, getRightCondition, interestDetail, rightholderGroups } from '@blockframes/model';
 import { createModalData } from '@blockframes/ui/global-modal/global-modal.component';
 import { DetailedGroupComponent } from '@blockframes/ui/detail-modal/detailed.component';
-import { filter, map, switchMap } from 'rxjs';
+import { combineLatest, filter, map, switchMap } from 'rxjs';
 import { DashboardWaterfallShellComponent } from '@blockframes/waterfall/dashboard/shell/shell.component';
 
 @Component({
@@ -38,6 +38,16 @@ export class DocumentComponent {
 
   public showExpenseTypes$ = this.contract$.pipe(
     map(contract => rightholderGroups.withStatements.includes(contract.type))
+  );
+
+  public interests$ = combineLatest([this.rights$, this.shell.state$, this.contract$]).pipe(
+    map(([rights, state, contract]) => {
+      const allConditions = rights.map(right => getRightCondition(right)).filter(condition => !!condition).flat();
+      const interestCondition = allConditions.find(condition => condition.name === 'interest');
+      if (!interestCondition) return;
+      const payload = interestCondition.payload as ConditionInterest;
+      return interestDetail(contract.id, payload, state.waterfall.state);
+    })
   );
 
   constructor(
