+++ libs/waterfall/src/lib/components/interests-table/interest-table.component.html
@@ -1,29 +1,26 @@
-<ng-container *ngIf="interests$ | async as interests">
-  <h3>Interest Rate Details</h3>
-  <bf-table [source]="interests"  pagination="50">
-    <ng-template colRef="event" label="Event" let-event>
-      {{ event }}
-    </ng-template>
-    <ng-template colRef="invested" label="Amount invested" let-invested>
-      <price-per-currency [price]="invested | formatPair"></price-per-currency>
-    </ng-template>
-    <ng-template colRef="revenues" label="Revenues earned" let-revenues>
-      <price-per-currency [price]="revenues | formatPair"></price-per-currency>
-    </ng-template>
-    <ng-template colRef="date" label="Event date" let-date>
-      {{ date | date: 'MM/dd/yyyy' }}
-    </ng-template>
-    <ng-template colRef="amountOwed" label="Amount Owed" let-amountOwed>
-      <price-per-currency [price]="amountOwed | formatPair"></price-per-currency>
-    </ng-template>
-    <ng-template colRef="periodDays" label="Period Days" let-periodDays>
-      <span matTooltip="Since previous event">{{ periodDays }}</span>
-    </ng-template>
-    <ng-template colRef="periodInterests" label="Period interests" let-periodInterests>
-      <price-per-currency [price]="periodInterests | formatPair"></price-per-currency>
-    </ng-template>
-    <ng-template colRef="interestOwed" label="Period interests owed" let-interestOwed>
-      <price-per-currency [price]="interestOwed | formatPair"></price-per-currency>
-    </ng-template>
-  </bf-table>
-</ng-container>
+<bf-table [source]="interests"  pagination="50">
+  <ng-template colRef="event" label="Event" let-event>
+    {{ event }}
+  </ng-template>
+  <ng-template colRef="invested" label="Amount invested" let-invested>
+    <price-per-currency [price]="invested | formatPair"></price-per-currency>
+  </ng-template>
+  <ng-template colRef="revenues" label="Revenues earned" let-revenues>
+    <price-per-currency [price]="revenues | formatPair"></price-per-currency>
+  </ng-template>
+  <ng-template colRef="date" label="Event date" let-date>
+    {{ date | date: 'MM/dd/yyyy' }}
+  </ng-template>
+  <ng-template colRef="amountOwed" label="Amount Owed" let-amountOwed>
+    <price-per-currency [price]="amountOwed | formatPair"></price-per-currency>
+  </ng-template>
+  <ng-template colRef="periodDays" label="Period Days" let-periodDays>
+    <span matTooltip="Since previous event">{{ periodDays }}</span>
+  </ng-template>
+  <ng-template colRef="periodInterests" label="Period interests" let-periodInterests>
+    <price-per-currency [price]="periodInterests | formatPair"></price-per-currency>
+  </ng-template>
+  <ng-template colRef="interestOwed" label="Period interests owed" let-interestOwed>
+    <price-per-currency [price]="interestOwed | formatPair"></price-per-currency>
+  </ng-template>
+</bf-table>
