+++ libs/waterfall/src/lib/components/statement-producer/summary/summary.module.ts
@@ -14,6 +14,7 @@ import { StatementIncomeEditModule } from '../../statement-income-edit/statement
 import { ExpenseTypePipeModule } from '../../../pipes/expense-type.pipe';
 import { RightHolderNamePipeModule } from '../../../pipes/rightholder-name.pipe';
 import { StatementExpenseEditModule } from '../../statement-expense-edit/statement-expense-edit.module';
+import { InterestTableModule } from '@blockframes/waterfall/components/interests-table/interest-table.module';
 
 // Components
 import { StatementProducerSummaryComponent } from './summary.component';
@@ -43,6 +44,7 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
     StatementExpenseEditModule,
     ExpenseTypePipeModule,
     RightHolderNamePipeModule,
+    InterestTableModule,
 
     // Material
     MatExpansionModule,
